Make widget clickable and do a small refactor of the updater
