+++ app/src/main/java/widget/MissionWidgetUpdater.kt
@@ -1,0 +1,46 @@
+package widget
+
+import android.content.Context
+import api.fetchData
+import data.MissionInfoEntry
+import kotlinx.coroutines.runBlocking
+import tools.formatMissionData
+import user.preferences.PreferencesDatastore
+import java.time.Instant
+
+class MissionWidgetUpdater {
+    fun updateMissions(context: Context) {
+        runBlocking {
+            val preferences = PreferencesDatastore(context)
+            var preferencesMissionData = preferences.getMissionInfo()
+            val prefEid = preferences.getEid()
+
+            try {
+                if (prefEid.isNotBlank()) {
+                    // Only make an api call if:
+                    // preferencesMissionData is has less than 3 active missions, meaning all active missions haven't been saved OR
+                    // preferencesMissionData has complete missions, meaning we need to fetch new active missions
+                    if (preferencesMissionData.size < 3 || anyMissionsComplete(
+                            preferencesMissionData
+                        )
+                    ) {
+                        val missionInfo = fetchData(prefEid)
+                        preferencesMissionData = formatMissionData(missionInfo)
+                    }
+                    preferences.saveMissionInfo(preferencesMissionData)
+                    MissionWidgetDataStore().setMissionInfo(context, preferencesMissionData)
+
+                    MissionWidgetDataStore().setEid(context, prefEid)
+                }
+            } catch (e: Exception) {
+                throw e
+            }
+        }
+    }
+
+    private fun anyMissionsComplete(missions: List<MissionInfoEntry>): Boolean {
+        return missions.any { mission ->
+            mission.secondsRemaining - (Instant.now().epochSecond - mission.date) <= 0
+        }
+    }
+}
