+++ app/src/main/res/xml/mission_widget.xml
@@ -3,6 +3,6 @@
     android:description="@string/app_name"
     android:minWidth="300dp"
     android:minHeight="50dp"
-    android:resizeMode="none"
+    android:resizeMode="horizontal|vertical"
     android:updatePeriodMillis="0"
     android:widgetCategory="home_screen" />
